Server java
I added a php file to connect to the server for uploads. I then went and added comments to the java script functions for better understanding. Then I added a function that has a try catch to prevent crashing when it request an object. After we added an if statement that activates the function when httpRequest is not false. Then for safety when a user spams the buttons more than once it will kill the extra clicks. After the abort is done then we set up an open to target a resource.