Visual data
We added a for loop with a lop of if else statements. the for loop loops through the days and the data and applies it to the data per data table. We also added some math equation to the code to round out the data sent for the weather percentage.